Address memory diagnostics review feedback
