+++ Sources/CmuxTopProcessSnapshotCache.swift
@@ -1,0 +1,37 @@
+import Foundation
+import os
+
+private nonisolated struct CmuxTopProcessSnapshotCacheState {
+    var snapshot: CmuxTopProcessSnapshot?
+    var includeProcessDetails = false
+}
+
+private nonisolated let cmuxTopProcessSnapshotCache = OSAllocatedUnfairLock(
+    initialState: CmuxTopProcessSnapshotCacheState()
+)
+
+nonisolated extension CmuxTopProcessSnapshot {
+    static func captureCached(
+        includeProcessDetails: Bool = false,
+        maximumAge: TimeInterval
+    ) -> CmuxTopProcessSnapshot {
+        let now = Date()
+        if let cached = cmuxTopProcessSnapshotCache.withLock({ state -> CmuxTopProcessSnapshot? in
+            guard state.includeProcessDetails == includeProcessDetails,
+                  let snapshot = state.snapshot,
+                  now.timeIntervalSince(snapshot.sampledAt) <= maximumAge else {
+                return nil
+            }
+            return snapshot
+        }) {
+            return cached
+        }
+
+        let snapshot = capture(includeProcessDetails: includeProcessDetails)
+        cmuxTopProcessSnapshotCache.withLock { state in
+            state.snapshot = snapshot
+            state.includeProcessDetails = includeProcessDetails
+        }
+        return snapshot
+    }
+}
