+++ Sources/CmuxTopSnapshot.swift
@@ -1,42 +1,4 @@
 import Foundation
-import Darwin
-import os
-
-private nonisolated let cmuxTopPIDPathBufferSize = 4096
-private nonisolated let cmuxTopMemoryDiagnosticDefaultGroupLimit = 12
-
-nonisolated struct CmuxTopProcessAttribution: Hashable, Sendable {
-    let workspaceID: UUID?
-    let workspaceRef: String?
-    let paneID: UUID?
-    let paneRef: String?
-    let surfaceID: UUID?
-    let surfaceRef: String?
-    let surfaceType: String?
-    let reason: String
-
-    func payload() -> [String: Any] {
-        [
-            "workspace_id": workspaceID?.uuidString as Any? ?? NSNull(),
-            "workspace_ref": workspaceRef as Any? ?? NSNull(),
-            "pane_id": paneID?.uuidString as Any? ?? NSNull(),
-            "pane_ref": paneRef as Any? ?? NSNull(),
-            "surface_id": surfaceID?.uuidString as Any? ?? NSNull(),
-            "surface_ref": surfaceRef as Any? ?? NSNull(),
-            "surface_type": surfaceType as Any? ?? NSNull(),
-            "reason": reason
-        ]
-    }
-}
-
-private nonisolated struct CmuxTopProcessSnapshotCacheState {
-    var snapshot: CmuxTopProcessSnapshot?
-    var includeProcessDetails = false
-}
-
-private nonisolated let cmuxTopProcessSnapshotCache = OSAllocatedUnfairLock(
-    initialState: CmuxTopProcessSnapshotCacheState()
-)
 
 nonisolated struct CmuxTopResourceSummary: Sendable {
     var cpuPercent: Double = 0
@@ -169,7 +131,7 @@ nonisolated struct CmuxTopProcessScope: Sendable {
 nonisolated final class CmuxTopProcessSnapshot: @unchecked Sendable {
     let sampledAt: Date
     private let includesProcessDetails: Bool
-    private let processesByPID: [Int: CmuxTopProcessInfo]
+    let processesByPID: [Int: CmuxTopProcessInfo]
     private let childrenByParentPID: [Int: [Int]]
     private let pidsByTTYDevice: [Int64: [Int]]
     private let pidsByCMUXSurfaceID: [UUID: [Int]]
@@ -181,30 +143,6 @@ nonisolated final class CmuxTopProcessSnapshot: @unchecked Sendable {
             sampledAt: Date(),
             includesProcessDetails: includeProcessDetails
         )
-    }
-
-    static func captureCached(
-        includeProcessDetails: Bool = false,
-        maximumAge: TimeInterval
-    ) -> CmuxTopProcessSnapshot {
-        let now = Date()
-        if let cached = cmuxTopProcessSnapshotCache.withLock({ state -> CmuxTopProcessSnapshot? in
-            guard state.includeProcessDetails == includeProcessDetails,
-                  let snapshot = state.snapshot,
-                  now.timeIntervalSince(snapshot.sampledAt) <= maximumAge else {
-                return nil
-            }
-            return snapshot
-        }) {
-            return cached
-        }
-
-        let snapshot = capture(includeProcessDetails: includeProcessDetails)
-        cmuxTopProcessSnapshotCache.withLock { state in
-            state.snapshot = snapshot
-            state.includeProcessDetails = includeProcessDetails
-        }
-        return snapshot
     }
 
     init(
@@ -325,50 +263,6 @@ nonisolated final class CmuxTopProcessSnapshot: @unchecked Sendable {
             stack.append(contentsOf: childrenByParentPID[pid] ?? [])
         }
         return result
-    }
-
-    func memoryDiagnosticPayload(
-        appPID: Int = Int(Darwin.getpid()),
-        topGroupLimit: Int = cmuxTopMemoryDiagnosticDefaultGroupLimit,
-        attributionByPID: [Int: CmuxTopProcessAttribution] = [:]
-    ) -> [String: Any] {
-        let appResources = summaryPayload(for: [appPID], rootPIDs: [appPID])
-        let appProcess = processesByPID[appPID]
-        let childPIDs = descendantPIDs(rootPID: appPID, includeRoot: false)
-            .filter { processesByPID[$0] != nil }
-        let childSummary = summary(for: childPIDs)
-        let groups = memoryDiagnosticGroups(
-            for: childPIDs,
-            topGroupLimit: topGroupLimit,
-            attributionByPID: attributionByPID
-        )
-        let topGroup = groups.first
-
-        return [
-            "sampled_at": ISO8601DateFormatter().string(from: sampledAt),
-            "app": [
-                "pid": appPID,
-                "name": appProcess?.name ?? "cmux",
-                "path": appProcess?.path as Any? ?? NSNull(),
-                "resources": appResources,
-                "physical_footprint_bytes": appProcess?.memoryBytes ?? 0,
-                "resident_bytes": appProcess?.residentBytes ?? 0,
-                "memory_source": appProcess?.memorySource.rawValue ?? CmuxTopProcessMemorySource.unavailable.rawValue,
-                "resident_memory_source": appProcess?.residentMemorySource.rawValue ?? CmuxTopProcessMemorySource.unavailable.rawValue
-            ] as [String: Any],
-            "children": [
-                "root_pid": appPID,
-                "recursive_rss_bytes": childSummary.residentBytes,
-                "process_count": childSummary.processCount,
-                "pids": childSummary.pids,
-                "groups": groups
-            ] as [String: Any],
-            "summary": memoryDiagnosticSummaryText(
-                appFootprintBytes: appProcess?.memoryBytes ?? 0,
-                childRSSBytes: childSummary.residentBytes,
-                topGroup: topGroup
-            )
-        ]
     }
 
     func summaryPayload(for pids: Set<Int>, rootPIDs: Set<Int> = []) -> [String: Any] {
@@ -707,404 +601,10 @@ nonisolated final class CmuxTopProcessSnapshot: @unchecked Sendable {
         return "\(process?.name ?? ""):\(pid)"
     }
 
-    private struct MemoryDiagnosticGroupAccumulator {
-        let id: String
-        let name: String
-        var rssBytes: Int64 = 0
-        var processIDs: [Int] = []
-        var attributions: [CmuxTopProcessAttribution: MemoryDiagnosticAttributionAccumulator] = [:]
-
-        mutating func append(
-            process: CmuxTopProcessInfo,
-            attribution: CmuxTopProcessAttribution?
-        ) {
-            rssBytes = CmuxTopProcessSnapshot.clampedAdd(rssBytes, process.residentBytes)
-            processIDs.append(process.pid)
-            guard let attribution else { return }
-            if attributions[attribution] == nil {
-                attributions[attribution] = MemoryDiagnosticAttributionAccumulator(attribution: attribution)
-            }
-            attributions[attribution]?.append(process: process)
-        }
-
-        func payload() -> [String: Any] {
-            let sortedProcessIDs = processIDs.sorted()
-            let attributionPayloads = attributions.values
-                .sorted {
-                    if $0.rssBytes != $1.rssBytes {
-                        return $0.rssBytes > $1.rssBytes
-                    }
-                    return $0.displayKey < $1.displayKey
-                }
-                .map { $0.payload() }
-            let topAttribution: Any = attributionPayloads.first.map { $0 as Any } ?? NSNull()
-            return [
-                "id": id,
-                "name": name,
-                "rss_bytes": rssBytes,
-                "resident_bytes": rssBytes,
-                "process_count": sortedProcessIDs.count,
-                "pids": sortedProcessIDs,
-                "top_attribution": topAttribution,
-                "attributions": attributionPayloads
-            ]
-        }
-    }
-
-    private struct MemoryDiagnosticAttributionAccumulator {
-        let attribution: CmuxTopProcessAttribution
-        var rssBytes: Int64 = 0
-        var processIDs: [Int] = []
-
-        var displayKey: String {
-            [
-                attribution.workspaceRef,
-                attribution.paneRef,
-                attribution.surfaceRef,
-                attribution.workspaceID?.uuidString,
-                attribution.paneID?.uuidString,
-                attribution.surfaceID?.uuidString
-            ]
-                .compactMap { $0 }
-                .joined(separator: "/")
-        }
-
-        mutating func append(process: CmuxTopProcessInfo) {
-            rssBytes = CmuxTopProcessSnapshot.clampedAdd(rssBytes, process.residentBytes)
-            processIDs.append(process.pid)
-        }
-
-        func payload() -> [String: Any] {
-            var payload = attribution.payload()
-            let sortedProcessIDs = processIDs.sorted()
-            payload["rss_bytes"] = rssBytes
-            payload["resident_bytes"] = rssBytes
-            payload["process_count"] = sortedProcessIDs.count
-            payload["pids"] = sortedProcessIDs
-            return payload
-        }
-    }
-
-    private func memoryDiagnosticGroups(
-        for pids: Set<Int>,
-        topGroupLimit: Int,
-        attributionByPID: [Int: CmuxTopProcessAttribution]
-    ) -> [[String: Any]] {
-        var groups: [String: MemoryDiagnosticGroupAccumulator] = [:]
-        for pid in pids.sorted() {
-            guard let process = processesByPID[pid] else { continue }
-            let name = process.name.trimmingCharacters(in: .whitespacesAndNewlines)
-            let displayName = name.isEmpty ? "pid-\(pid)" : name
-            let key = displayName.lowercased()
-            if groups[key] == nil {
-                groups[key] = MemoryDiagnosticGroupAccumulator(id: key, name: displayName)
-            }
-            groups[key]?.append(
-                process: process,
-                attribution: attributionByPID[pid] ?? nearestCMUXAttribution(for: pid)
-            )
-        }
-
-        let limit = max(1, topGroupLimit)
-        return groups.values
-            .sorted {
-                if $0.rssBytes != $1.rssBytes {
-                    return $0.rssBytes > $1.rssBytes
-                }
-                return $0.name.localizedStandardCompare($1.name) == .orderedAscending
-            }
-            .prefix(limit)
-            .map { $0.payload() }
-    }
-
-    private func nearestCMUXAttribution(for pid: Int) -> CmuxTopProcessAttribution? {
-        var visited: Set<Int> = []
-        var currentPID = pid
-        while currentPID > 0, visited.insert(currentPID).inserted {
-            guard let process = processesByPID[currentPID] else { return nil }
-            if process.cmuxWorkspaceID != nil || process.cmuxSurfaceID != nil {
-                return CmuxTopProcessAttribution(
-                    workspaceID: process.cmuxWorkspaceID,
-                    workspaceRef: nil,
-                    paneID: nil,
-                    paneRef: nil,
-                    surfaceID: process.cmuxSurfaceID,
-                    surfaceRef: nil,
-                    surfaceType: "terminal",
-                    reason: process.cmuxAttributionReason ?? "cmux-process-scope"
-                )
-            }
-            currentPID = process.parentPID
-        }
-        return nil
-    }
-
-    private func memoryDiagnosticSummaryText(
-        appFootprintBytes: Int64,
-        childRSSBytes: Int64,
-        topGroup: [String: Any]?
-    ) -> String {
-        var summary = String.localizedStringWithFormat(
-            String(localized: "memoryDiagnostic.summary.base", defaultValue: "%@ app footprint + %@ child RSS"),
-            Self.formatDiagnosticBytes(appFootprintBytes),
-            Self.formatDiagnosticBytes(childRSSBytes)
-        )
-        guard let topGroup,
-              let name = topGroup["name"] as? String,
-              let rssBytes = topGroup["rss_bytes"] as? Int64 ?? (topGroup["rss_bytes"] as? NSNumber)?.int64Value else {
-            return summary
-        }
-
-        summary += String.localizedStringWithFormat(
-            String(localized: "memoryDiagnostic.summary.topGroup", defaultValue: "; top child group: %@ %@"),
-            name,
-            Self.formatDiagnosticBytes(rssBytes)
-        )
-        if let attribution = topGroup["top_attribution"] as? [String: Any],
-           let workspace = attribution["workspace_ref"] as? String ?? attribution["workspace_id"] as? String,
-           !workspace.isEmpty {
-            summary += String.localizedStringWithFormat(
-                String(localized: "memoryDiagnostic.summary.workspace", defaultValue: " from workspace %@"),
-                workspace
-            )
-        }
-        return summary
-    }
-
-    private static func formatDiagnosticBytes(_ bytes: Int64) -> String {
-        let formatter = ByteCountFormatter()
-        formatter.countStyle = .memory
-        formatter.allowedUnits = .useAll
-        return formatter.string(fromByteCount: max(0, bytes))
-    }
-
-    private static func allProcesses(includeProcessDetails: Bool) -> [CmuxTopProcessInfo] {
-        let sampledProcesses = allBSDProcesses()
-        guard !sampledProcesses.isEmpty else { return [] }
-
-        var scopeKeyByPID: [Int: CmuxTopProcessScopeCacheKey] = [:]
-        scopeKeyByPID.reserveCapacity(sampledProcesses.count)
-        for process in sampledProcesses {
-            scopeKeyByPID[Int(process.pbi_pid)] = scopeCacheKey(from: process)
-        }
-        let activeScopeKeys = Set(scopeKeyByPID.values)
-        var parentScopeKeys: [CmuxTopProcessScopeCacheKey: CmuxTopProcessScopeCacheKey] = [:]
-        parentScopeKeys.reserveCapacity(sampledProcesses.count)
-        for process in sampledProcesses {
-            let key = scopeCacheKey(from: process)
-            let parentPID = Int(process.pbi_ppid)
-            guard let parentKey = scopeKeyByPID[parentPID] else { continue }
-            parentScopeKeys[key] = parentKey
-        }
-        let sampledAtNanoseconds = cpuSampleClockNanoseconds()
-        var currentCPUSamples: [CmuxTopProcessScopeCacheKey: CmuxTopProcessCPUSample] = [:]
-        var processRecords: [(info: CmuxTopProcessInfo, cpuSampleKey: CmuxTopProcessScopeCacheKey?)] = []
-        processRecords.reserveCapacity(sampledProcesses.count)
-        for process in sampledProcesses {
-            guard let processRecord = processInfo(
-                from: process,
-                includeProcessDetails: includeProcessDetails,
-                sampledAtNanoseconds: sampledAtNanoseconds,
-                currentCPUSamples: &currentCPUSamples
-            ) else {
-                continue
-            }
-            processRecords.append(processRecord)
-        }
-        let cpuPercentages = cpuPercentages(
-            for: currentCPUSamples,
-            activeKeys: activeScopeKeys,
-            parentKeysByKey: parentScopeKeys,
-            sampledAtNanoseconds: sampledAtNanoseconds
-        )
-        for index in processRecords.indices {
-            guard let key = processRecords[index].cpuSampleKey,
-                  let cpuPercent = cpuPercentages[key] else { continue }
-            processRecords[index].info.cpuPercent = cpuPercent
-        }
-        pruneCMUXScopeCache(activeKeys: activeScopeKeys)
-        return processRecords.map(\.info)
-    }
-
-    private static func processInfo(
-        from bsdInfo: proc_bsdinfo,
-        includeProcessDetails: Bool,
-        sampledAtNanoseconds: UInt64,
-        currentCPUSamples: inout [CmuxTopProcessScopeCacheKey: CmuxTopProcessCPUSample]
-    ) -> (info: CmuxTopProcessInfo, cpuSampleKey: CmuxTopProcessScopeCacheKey?)? {
-        let pid = Int(bsdInfo.pbi_pid)
-        guard pid > 0 else { return nil }
-
-        let taskInfo = taskInfo(for: pid)
-        let resourceUsage = resourceUsage(for: pid)
-        let cacheKey = scopeCacheKey(from: bsdInfo)
-        let fallbackName = fixedString(bsdInfo.pbi_comm)
-        let name = includeProcessDetails ? processName(pid: pid, fallback: fallbackName) : fallbackName
-        let path = includeProcessDetails ? processPath(pid: pid) : nil
-        let rawTTY = Int64(bsdInfo.e_tdev)
-        let ttyDevice = rawTTY > 0 ? rawTTY : nil
-        let cmuxScope = cachedCMUXScope(for: pid, cacheKey: cacheKey)
-        let rawProcessGroupID = Int(bsdInfo.pbi_pgid)
-        let processGroupID = rawProcessGroupID > 0 ? rawProcessGroupID : nil
-        let rawTerminalProcessGroupID = Int(bsdInfo.e_tpgid)
-        let terminalProcessGroupID = rawTerminalProcessGroupID > 0 ? rawTerminalProcessGroupID : nil
-        let memoryBytes: Int64
-        let memorySource: CmuxTopProcessMemorySource
-        if let resourceUsage {
-            memoryBytes = int64Clamped(resourceUsage.ri_phys_footprint)
-            memorySource = .physicalFootprint
-        } else if let taskInfo {
-            memoryBytes = int64Clamped(taskInfo.pti_resident_size)
-            memorySource = .residentSize
-        } else {
-            memoryBytes = 0
-            memorySource = .unavailable
-        }
-        let residentBytes: Int64
-        let residentMemorySource: CmuxTopProcessMemorySource
-        if let taskInfo {
-            residentBytes = int64Clamped(taskInfo.pti_resident_size)
-            residentMemorySource = .residentSize
-        } else if let resourceUsage {
-            residentBytes = int64Clamped(resourceUsage.ri_resident_size)
-            residentMemorySource = .rusageResidentSize
-        } else {
-            residentBytes = 0
-            residentMemorySource = .unavailable
-        }
-        let cpuSampleKey: CmuxTopProcessScopeCacheKey?
-        if let taskInfo {
-            let currentCPUSample = cpuSample(from: taskInfo, sampledAtNanoseconds: sampledAtNanoseconds)
-            currentCPUSamples[cacheKey] = currentCPUSample
-            cpuSampleKey = cacheKey
-        } else {
-            cpuSampleKey = nil
-        }
-
-        return (CmuxTopProcessInfo(
-            pid: pid,
-            parentPID: Int(bsdInfo.pbi_ppid),
-            name: name.isEmpty ? "pid-\(pid)" : name,
-            path: path,
-            ttyDevice: ttyDevice,
-            cmuxWorkspaceID: cmuxScope?.workspaceID,
-            cmuxSurfaceID: cmuxScope?.surfaceID,
-            cmuxAttributionReason: cmuxScope?.attributionReason,
-            processGroupID: processGroupID,
-            terminalProcessGroupID: terminalProcessGroupID,
-            cpuPercent: 0,
-            memoryBytes: memoryBytes,
-            memorySource: memorySource,
-            residentBytes: residentBytes,
-            residentMemorySource: residentMemorySource,
-            virtualBytes: int64Clamped(taskInfo?.pti_virtual_size ?? 0),
-            threadCount: Int(taskInfo?.pti_threadnum ?? 0)
-        ), cpuSampleKey)
-    }
-
-    private static func allBSDProcesses() -> [proc_bsdinfo] {
-        let pidStride = MemoryLayout<pid_t>.stride
-        for _ in 0..<3 {
-            let pidCount = Int(proc_listallpids(nil, 0))
-            guard pidCount > 0 else { return [] }
-            var pids = Array(repeating: pid_t(), count: max(1, pidCount + 32))
-            let returnedCount = pids.withUnsafeMutableBufferPointer { buffer in
-                proc_listallpids(buffer.baseAddress, Int32(buffer.count * pidStride))
-            }
-            guard returnedCount >= 0 else { return [] }
-            let count = min(pids.count, Int(returnedCount))
-            if count < pids.count {
-                return pids.prefix(count).compactMap { pid in
-                    guard pid > 0 else { return nil }
-                    return bsdInfo(for: Int(pid))
-                }
-            }
-        }
-        return []
-    }
-
-    private static func bsdInfo(for pid: Int) -> proc_bsdinfo? {
-        var info = proc_bsdinfo()
-        let expectedSize = MemoryLayout<proc_bsdinfo>.stride
-        let size = proc_pidinfo(pid_t(pid), PROC_PIDTBSDINFO, 0, &info, Int32(expectedSize))
-        return size == expectedSize ? info : nil
-    }
-
-    private static func deviceIdentifier(forTTYName ttyName: String) -> Int64? {
-        let trimmed = ttyName.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !trimmed.isEmpty, trimmed != "not a tty" else {
-            return nil
-        }
-
-        let path: String
-        if trimmed.hasPrefix("/dev/") {
-            path = trimmed
-        } else {
-            path = "/dev/\(trimmed)"
-        }
-
-        var statInfo = stat()
-        guard stat(path, &statInfo) == 0 else {
-            return nil
-        }
-        return Int64(statInfo.st_rdev)
-    }
-
-    private static func clampedAdd(_ lhs: Int64, _ rhs: Int64) -> Int64 {
+    static func clampedAdd(_ lhs: Int64, _ rhs: Int64) -> Int64 {
         if rhs > 0, lhs > Int64.max - rhs {
             return Int64.max
         }
         return lhs + rhs
-    }
-
-    private static func taskInfo(for pid: Int) -> proc_taskinfo? {
-        var info = proc_taskinfo()
-        let expectedSize = MemoryLayout<proc_taskinfo>.stride
-        let size = proc_pidinfo(pid_t(pid), PROC_PIDTASKINFO, 0, &info, Int32(expectedSize))
-        return size == expectedSize ? info : nil
-    }
-
-    private static func resourceUsage(for pid: Int) -> rusage_info_v4? {
-        var info = rusage_info_v4()
-        let result = withUnsafeMutableBytes(of: &info) { rawBuffer -> Int32 in
-            guard let baseAddress = rawBuffer.baseAddress else { return -1 }
-            // proc_pid_rusage imports as rusage_info_t *; callers pass the concrete
-            // rusage struct address cast to that opaque buffer type.
-            let buffer = baseAddress.assumingMemoryBound(to: rusage_info_t?.self)
-            return proc_pid_rusage(
-                pid_t(pid),
-                RUSAGE_INFO_V4,
-                buffer
-            )
-        }
-        return result == 0 ? info : nil
-    }
-
-    private static func processName(pid: Int, fallback: String) -> String {
-        var buffer = [CChar](repeating: 0, count: Int(MAXCOMLEN + 1))
-        let length = proc_name(pid_t(pid), &buffer, UInt32(buffer.count))
-        guard length > 0 else { return fallback }
-        let name = String(cString: buffer).trimmingCharacters(in: .whitespacesAndNewlines)
-        return name.isEmpty ? fallback : name
-    }
-
-    private static func processPath(pid: Int) -> String? {
-        var buffer = [CChar](repeating: 0, count: cmuxTopPIDPathBufferSize)
-        let length = proc_pidpath(pid_t(pid), &buffer, UInt32(buffer.count))
-        guard length > 0 else { return nil }
-        let path = String(cString: buffer).trimmingCharacters(in: .whitespacesAndNewlines)
-        return path.isEmpty ? nil : path
-    }
-
-    private static func fixedString<T>(_ value: T) -> String {
-        withUnsafeBytes(of: value) { rawBuffer in
-            let endIndex = rawBuffer.firstIndex(of: 0) ?? rawBuffer.endIndex
-            return String(decoding: rawBuffer[..<endIndex], as: UTF8.self)
-                .trimmingCharacters(in: .whitespacesAndNewlines)
-        }
-    }
-
-    private static func int64Clamped(_ value: UInt64) -> Int64 {
-        value > UInt64(Int64.max) ? Int64.max : Int64(value)
     }
 }
